+++ boofuzz/connections/raw_l2_socket_connection.py
@@ -6,11 +6,14 @@ import sys
 
 from future.utils import raise_
 
-from . import base_socket_connection, exception
+from boofuzz import exception
+from boofuzz.connections import base_socket_connection
 
 
 class RawL2SocketConnection(base_socket_connection.BaseSocketConnection):
     """BaseSocketConnection implementation for use with Raw Layer 2 Sockets.
+
+    .. versionadded:: 0.2.0
 
     Args:
         interface (str): Hostname or IP adress of target system.