+++ boofuzz/connections/raw_l3_socket_connection.py
@@ -6,13 +6,17 @@ import sys
 
 from future.utils import raise_
 
-from . import base_socket_connection, exception
+from boofuzz import exception
+from boofuzz.connections import base_socket_connection
 
+ETH_P_ALL = 0x0003  # Ethernet protocol: Every packet, see Linux if_ether.h docs for more details.
 ETH_P_IP = 0x0800  # Ethernet protocol: Internet Protocol packet, see Linux <net/if_ether.h> docs for more details.
 
 
 class RawL3SocketConnection(base_socket_connection.BaseSocketConnection):
     """BaseSocketConnection implementation for use with Raw Layer 2 Sockets.
+
+    .. versionadded:: 0.2.0
 
     Args:
         interface (str): Interface to send and receive on.